move async race tests to their own file
